feat: user modeling, product removed, dto moved to dto folder
