--- src/main/java/com/tcc5/car_price_compare/domain/product/Product.java
@@ -1,28 +1,0 @@
-package com.tcc5.car_price_compare.domain.product;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Table(name = "product")
-@Entity(name = "product")
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode(of = "id")
-public class Product {
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private String id;
-
-    private String name;
-
-    private Integer price;
-
-    public Product(ProductRequestDTO data){
-        this.price = data.price();
-        this.name = data.name();
-    }
-}
