--- src/main/java/com/tcc5/car_price_compare/domain/product/ProductResponseDTO.java
@@ -1,7 +1,0 @@
-package com.tcc5.car_price_compare.domain.product;
-
-public record ProductResponseDTO(String id, String name, Integer price) {
-    public ProductResponseDTO(Product product){
-        this(product.getId(), product.getName(), product.getPrice());
-    }
-}
