--- src/main/java/com/tcc5/car_price_compare/domain/user/LoginResponseDTO.java
@@ -1,4 +1,0 @@
-package com.tcc5.car_price_compare.domain.user;
-
-public record LoginResponseDTO(String token) {
-}
