+++ src/main/java/com/tcc5/car_price_compare/domain/user/dto/AuthenticationDTO.java
@@ -1,4 +1,4 @@
-package com.tcc5.car_price_compare.domain.user;
+package com.tcc5.car_price_compare.domain.user.dto;
 
 public record AuthenticationDTO(String login, String password) {
 }
