+++ src/main/java/com/tcc5/car_price_compare/domain/user/dto/LoginResponseDTO.java
@@ -1,0 +1,4 @@
+package com.tcc5.car_price_compare.domain.user.dto;
+
+public record LoginResponseDTO(String token) {
+}
