+++ src/main/java/com/tcc5/car_price_compare/repositories/UserRepository.java
@@ -4,6 +4,8 @@ import com.tcc5.car_price_compare.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 
-public interface UserRepository extends JpaRepository<User, String> {
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
     UserDetails findByLogin(String login);
 }
